refactor(plugins): fix Supermemory React Doctor warnings (#970)
