+++ packages/plugins/supermemory/src/ui-browser.tsx
@@ -1,0 +1,66 @@
+/** @jsxRuntime classic */
+/** @jsx React.createElement */
+/** @jsxFrag React.Fragment */
+
+import { createCollection } from "./ui-collection";
+import type { Context, Profile } from "./ui-context";
+export function createBrowser(ctx: Context) {
+  const React = ctx.React;
+  const {
+    Button,
+    Select,
+    SelectTrigger,
+    SelectValue,
+    SelectContent,
+    SelectItem,
+  } = ctx.ui;
+  const Collection = createCollection(ctx);
+  return function Browser({ profiles }: { profiles: Profile[] }) {
+    const [agentId, setAgentId] = React.useState(profiles[0]?.id ?? "");
+    const [kind, setKind] = React.useState<"memory" | "knowledge">("memory");
+    return (
+      <>
+        <Select
+          onValueChange={(value) => setAgentId(value ?? "")}
+          value={agentId}
+        >
+          <SelectTrigger aria-label="Agent">
+            <SelectValue placeholder="Choose agent" />
+          </SelectTrigger>
+          <SelectContent>
+            {profiles.map((profile) => (
+              <SelectItem key={profile.id} value={profile.id}>
+                {profile.name}
+              </SelectItem>
+            ))}
+          </SelectContent>
+        </Select>
+        <div aria-label="Collection" className="sm-row" role="group">
+          <Button
+            aria-pressed={kind === "memory"}
+            onClick={() => setKind("memory")}
+            variant={kind === "memory" ? "default" : "ghost"}
+          >
+            Memory
+          </Button>
+          <Button
+            aria-pressed={kind === "knowledge"}
+            onClick={() => setKind("knowledge")}
+            variant={kind === "knowledge" ? "default" : "ghost"}
+          >
+            Knowledge
+          </Button>
+        </div>
+        {agentId ? (
+          <Collection
+            agentId={agentId}
+            key={`${agentId}:${kind}`}
+            kind={kind}
+          />
+        ) : (
+          <p>No agents available</p>
+        )}
+      </>
+    );
+  };
+}
