+++ packages/plugins/supermemory/src/ui-page.tsx
@@ -1,0 +1,111 @@
+/** @jsxRuntime classic */
+/** @jsx React.createElement */
+/** @jsxFrag React.Fragment */
+
+import { createBrowser } from "./ui-browser";
+import { type Context, errorText, type Profile } from "./ui-context";
+import { createSettings } from "./ui-settings";
+
+function usePage(ctx: Context) {
+  const React = ctx.React;
+  const [profiles, setProfiles] = React.useState<Profile[]>([]);
+  const [canConfigure, setCanConfigure] = React.useState(false);
+  const [configured, setConfigured] = React.useState(false);
+  const [settings, setSettings] = React.useState(false);
+  const [error, setError] = React.useState("");
+  const [loading, setLoading] = React.useState(true);
+  React.useEffect(() => {
+    let alive = true;
+    ctx.host
+      .call("profiles")
+      .then((value) => {
+        if (!alive || ctx.signal.aborted) {
+          return;
+        }
+        const result = value as {
+          profiles: Profile[];
+          configured: boolean;
+          canConfigure: boolean;
+        };
+        setProfiles(result.profiles);
+        setConfigured(result.configured);
+        setCanConfigure(result.canConfigure);
+      })
+      .catch((reason) => {
+        if (alive) {
+          setError(errorText(reason));
+        }
+      })
+      .finally(() => {
+        if (alive) {
+          setLoading(false);
+        }
+      });
+    return () => {
+      alive = false;
+    };
+  }, []);
+  return {
+    canConfigure,
+    configured,
+    error,
+    loading,
+    profiles,
+    setConfigured,
+    setSettings,
+    settings,
+  };
+}
+export function createPage(ctx: Context) {
+  const React = ctx.React;
+  const { Button } = ctx.ui;
+  const Settings = createSettings(ctx);
+  const Browser = createBrowser(ctx);
+  function Page() {
+    const {
+      profiles,
+      canConfigure,
+      configured,
+      setConfigured,
+      settings,
+      setSettings,
+      error,
+      loading,
+    } = usePage(ctx);
+    return (
+      <section className="sm-page sm-stack">
+        <header className="sm-row">
+          <h1 className="sm-title">Supermemory</h1>
+          {configured && <span>Configured</span>}
+          {canConfigure && (
+            <Button onClick={() => setSettings(true)} variant="outline">
+              Settings
+            </Button>
+          )}
+        </header>
+        {error && <p role="alert">{error}</p>}
+        {loading ? (
+          <p role="status">Loading…</p>
+        ) : configured ? (
+          <Browser profiles={profiles} />
+        ) : (
+          <p>
+            {canConfigure
+              ? "Connect your Supermemory server in Settings."
+              : "Ask an admin to connect Supermemory."}
+          </p>
+        )}
+        {settings && (
+          <Settings
+            close={() => setSettings(false)}
+            saved={() => {
+              setConfigured(true);
+              setSettings(false);
+            }}
+          />
+        )}
+      </section>
+    );
+  }
+  return Page;
+}
